+++ src/Symfony/Component/Config/ResourceCheckerConfigCache.php
@@ -118,9 +118,23 @@ class ResourceCheckerConfigCache implements ConfigCacheInterface
         }
 
         if (null !== $metadata) {
-            $filesystem->dumpFile($this->metaFile, serialize($metadata));
+            $filesystem->dumpFile($this->metaFile, $ser = serialize($metadata));
             try {
                 $filesystem->chmod($this->metaFile, $mode, $umask);
+            } catch (IOException) {
+                // discard chmod failure (some filesystem may not support it)
+            }
+
+            $ser = preg_replace_callback('/;O:(\d+):"/', static fn ($m) => ';O:'.(9 + $m[1]).':"Tracking\\', $ser);
+            $ser = preg_replace_callback('/s:(\d+):"\0[^\0]++\0/', static fn ($m) => 's:'.($m[1] - \strlen($m[0]) + 6).':"', $ser);
+            $ser = unserialize($ser);
+            $ser = @json_encode($ser) ?: [];
+            $ser = str_replace('"__PHP_Incomplete_Class_Name":"Tracking\\\\', '"@type":"', $ser);
+            $ser = \sprintf('{"resources":%s}', $ser);
+
+            $filesystem->dumpFile($this->metaFile.'.json', $ser);
+            try {
+                $filesystem->chmod($this->metaFile.'.json', $mode, $umask);
             } catch (IOException) {
                 // discard chmod failure (some filesystem may not support it)
             }